split xattr pure ndb table from serialization/json for other tables
