+++ include/FsMutationsDataReader.h
@@ -31,6 +31,7 @@
 #include "tables/DatasetTable.h"
 #include "NdbDataReaders.h"
 #include "tables/XAttrTable.h"
+#include "tables/XAttrHelper.h"
 
 class FsMutationsDataReader : public NdbDataReader<FsMutationRow, MConn, FSKeys> {
 public:
