+++ src/Reindexer.cpp
@@ -29,6 +29,7 @@
 #include "tables/SchemabasedMetadataTable.h"
 #include "tables/SchemalessMetadataTable.h"
 #include "tables/XAttrTable.h"
+#include "tables/XAttrHelper.h"
 
 struct DatasetInodes {
   Int64 mDatasetId;
@@ -182,7 +183,7 @@ void Reindexer::run() {
     for(XAttrVec::iterator xit = xattrs.begin(); xit != xattrs.end(); ++xit){
       XAttrRow xAttrRow = *xit;
       if(xAttrRow.mInodeId == inodeId){
-        bulk.mJSON += xAttrRow.to_upsert_json();
+        bulk.mJSON += XAttrHelper::to_upsert_json(xAttrRow);
       }else{
         LOG_WARN("XAttrs doesn't exists for ["
                      << inodeId << "] - " << xAttrRow.to_string());
